scimage: fix bug in setup_environ()
Updated setup_environ() to check whether /sbin/initctl is already a
symlink or not before applying the initctl hack for apt-get. This avoids
backing-up a symlink and destroying the real back-up in the case that
setup_environ() is called multiple times without a cleanup() in between.

+++ utils/scimage.py
@@ -517,7 +517,7 @@ def setup_environ():
     num_cpus = multiprocessing.cpu_count()
     os.environ['MAKEFLAGS'] = '-j%d' % (num_cpus + 1)
     os.environ['DEBIAN_FRONTEND'] = "noninteractive"
-    if os.path.isfile('/sbin/initctl'):
+    if os.path.isfile('/sbin/initctl') and not os.path.islink('/sbin/initctl'):
         run_command('mv /sbin/initctl /sbin/initctl.bak')
         run_command('ln -s /bin/true /sbin/initctl')
